requirements.txt
fastapi
uvicorn

psycopg2-binary

langgraph
langchain

transformers
torch
accelerate

python-dotenv
pydantic

matplotlib
pandas
